dpo_requirements.txt
absl-py==2.2.2
accelerate==1.3.0
aenum==3.1.15
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-generator==1.10
async-lru==2.0.4
attrs==24.3.0
autoawq==0.2.9
babel==2.16.0
bash_kernel==0.9.3
beautifulsoup4==4.12.3
bitsandbytes==0.45.5
bleach==6.2.0
cachetools==5.5.0
certifi==2024.12.14
certipy==0.2.1
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
compressed-tensors==0.9.4
contourpy==1.3.1
croniter==1.4.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.10
dataclasses-json==0.6.7
DataProperty==1.1.0
datasets==3.2.0
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.9
docker-pycreds==0.4.0
drawsvg==2.4.0
einops==0.8.1
evaluate==0.4.3
executing==2.1.0
faiss-gpu-cu12==1.10.0
fastjsonschema==2.21.1
filelock==3.17.0
fonttools==4.55.8
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
gspread==6.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.31.2
idna==3.10
import-ipynb==0.1.4
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter-server-mathjax==0.2.6
jupyter-telemetry==0.1.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_scheduler==2.7.1
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterhub==4.1.5
jupyterlab==4.4.2
jupyterlab_git==0.51.1
jupyterlab_myst==2.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.59
langchain-huggingface==0.2.0
langchain-text-splitters==0.3.8
langsmith==0.3.42
latex2sympy2_extended==1.10.1
-e git+https://github.com/eric11eca/lighteval-epfl-mnlp.git@8229f7ee4ac59b15a08f8021c2286910e1169b45#egg=lighteval
lxml==5.3.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
mdurl==0.1.2
mistral_common==1.5.6
mistune==3.1.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy_extensions==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbdime==4.0.2
nbformat==5.10.4
nbgitpuller==1.2.2
nbgitpuller-link==0.2.5
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
notebook==7.4.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
opencv-python==4.11.0.86
optimum-quanto==0.2.7
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pamela==1.2.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathvalidate==3.2.3
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
playwright==1.49.1
portalocker==3.1.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyee==12.0.0
Pygments==2.19.1
pyparsing==3.2.1
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.2.1
python-on-whales==0.71.0
pytz==2023.3
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
sacrebleu==2.5.1
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.14.0
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.4.1
sentencepiece==0.2.0
sentry-sdk==2.29.1
setproctitle==1.3.6
setuptools==70.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
stack-data==0.6.3
sympy==1.14.0
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tenacity==9.1.2
termcolor==2.3.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.0
torch==2.7.0
torch-geometric==2.6.1
torchaudio==2.5.1
torchsummary==1.5.1
torchvision==0.22.0+cu126
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.3.0
trl==0.17.0
typepy==1.3.4
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
validators==0.34.0
virtualenv==20.29.1
voila==0.5.7
wandb==0.19.11
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
wheel==0.45.1
widgetsnbextension==4.0.13
wordcloud==1.9.4
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
